Re-organized files and correct classes
