+++ entities/common.js
@@ -1,0 +1,6 @@
+import { Ghost } from './ghost.js';
+import { Player } from './player.js';
+import { Boundary } from './boundary.js';
+import { Pellet } from './pellet.js';
+
+export default { Ghost, Player, Boundary, Pellet };
