+++ maps/mapList.js
@@ -1,0 +1,19 @@
+export const maps = [];
+
+const map = [
+	['1', '-', '-', '-', '-', '-', '-', '-', '-', '-', '2'],
+	['|', '.,gred', '.', '.', '.', '.', '.', '.', '.', '.', '|'],
+	['|', '.', 'b', '.', '[', '7', ']', '.', 'b', '.', '|'],
+	['|', '.', '.', '.', '.', '_', '.', '.', '.', '.', '|'],
+	['|', '.', '[', ']', '.', 'p', '.', '[', ']', '.', '|'],
+	['|', '.', '.', '.', '.', '^', '.', '.', '.', '.', '|'],
+	['|', '.', 'b', '.', '[', '+', ']', '.', 'b', '.', '|'],
+	['|', '.', '.', '.', '.', '_', '.', '.', '.', '.', '|'],
+	['|', '.', '[', ']', '.', '.', '.', '[', ']', '.', '|'],
+	['|', '.', '.', '.', '.', '^', '.', '.', '.', '.', '|'],
+	['|', '.', 'b', '.', '[', '5', ']', '.', 'b', '.', '|'],
+	['|', '.', '.', '.', '.', '.', '.', '.', '.', '.,gpink', '|'],
+	['4', '-', '-', '-', '-', '-', '-', '-', '-', '-', '3'],
+];
+
+maps.push(map);
